Fix compiler warning
git-svn-id: https://asmotor.googlecode.com/svn/trunk@72 fe0e4675-b9e4-17b0-9a9e-715403563874

+++ xasm/common/lexer.c
@@ -626,7 +626,7 @@ void lex_RemoveString(char* pszName, int nToken)
 	
 	while(pToken)
 	{
-		if(pToken->Token == nToken
+		if(pToken->Token == (uint32_t)nToken
 		&& _stricmp(pToken->pszName, pszName) == 0)
 		{
 			list_Remove(*pHash, pToken);
